pyproject.toml
[project]
name = "pair-prgm-learn"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Wsine", email = "weizhy3@hotmail.com"},
]
dependencies = [
    "openai>=1.13.3",
    "Jinja2>=3.1.3",
    "python-dotenv>=1.0.1",
    "marko>=2.0.3",
    "mosspy>=1.0.9",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
package-type = "application"


[tool.pyright]
exclude = ["**/__pycache__"]
ignore = [".venv"]
venvPath = "./"
venv = ".venv"
strictParameterNoneValue = false

